Major update
Now it is possible to read files from Lab 3 (newest format) and use the plotting and analysis subroutines. The phases need to be specified as a constant offset in a file called "Phases.txt" in the same folder of the data (.CSV files), containing the rephasing and non-rephasing time-domain data (joined together).