Major Updates: Added 2D Gaussian Fit
Major updates to the 2D-IR analysis routine. The 2D Gaussian fit has been added, and the GUI and overall code has been revised and optimised.

A new major version of the GUI, combining the 2D-IR and pump-probe analysis routines is in active development. This version will use the App Designer MATLAB interface instead of the old-fashioned GUIDE.
This will take some time and effort, as the code is not directly portable, but will hopefully allow the two interfaces to be unified.