Added new chirp correction method
Added new automated chirp correction method based on fitting of a Gaussian-convolved step function and the coherent artifact (modelled as Gaussian+sum of derivatives).

The default t0 for this method is taken as the maximum of |d(DAbs)/dt| [the derivative of a step is a delta], which works quite well.